flask_auth_app/requirements.txt
quart
Flask-mail
werkzeug
requests
telethon
mysql-connector